+++ discovery-provider/integration_tests/tasks/test_index_trending_notification.py
@@ -4,7 +4,7 @@ from sqlalchemy import asc
 
 from integration_tests.utils import populate_mock_db
 from src.models.notifications.notification import Notification
-from src.queries.get_trending_tracks import make_trending_cache_key
+from src.queries.generate_unpopulated_trending import make_trending_cache_key
 from src.tasks.index_trending import (
     get_top_trending_to_notify,
     index_trending_notifications,
