+++ discovery-provider/src/queries/get_underground_trending.py
@@ -17,11 +17,8 @@ from src.premium_content.premium_content_constants import (
     SHOULD_TRENDING_EXCLUDE_COLLECTIBLE_GATED_TRACKS,
     SHOULD_TRENDING_EXCLUDE_PREMIUM_TRACKS,
 )
-from src.queries.get_trending_tracks import (
-    TRENDING_LIMIT,
-    TRENDING_TTL_SEC,
-    make_trending_cache_key,
-)
+from src.queries.generate_unpopulated_trending import make_trending_cache_key
+from src.queries.get_trending_tracks import TRENDING_LIMIT, TRENDING_TTL_SEC
 from src.queries.get_unpopulated_tracks import get_unpopulated_tracks
 from src.queries.query_helpers import (
     get_karma,
